Refactor HnStoryContentCard to remove unused imports and share functionality; add HnStoryMetaRow to HnStoryPage for enhanced story metadata display
